Rename rqd tests so they are picked up by pytest
